Move logic to a class, add constants and another test.
