+++ game-app/game-core/src/main/java/games/strategy/triplea/attachments/UnitAttachment.java
@@ -1,7 +1,6 @@
 package games.strategy.triplea.attachments;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.Iterables;
 import games.strategy.engine.data.Attachable;
 import games.strategy.engine.data.DefaultAttachment;
 import games.strategy.engine.data.DefaultNamed;
@@ -16,7 +15,6 @@ import games.strategy.engine.data.Unit;
 import games.strategy.engine.data.UnitType;
 import games.strategy.engine.data.UnitTypeList;
 import games.strategy.engine.data.gameparser.GameParseException;
-import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.triplea.Constants;
 import games.strategy.triplea.Properties;
 import games.strategy.triplea.delegate.Matches;
@@ -34,7 +32,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
@@ -42,7 +39,6 @@ import lombok.Value;
 import org.triplea.java.ChangeOnNextMajorRelease;
 import org.triplea.java.collections.CollectionUtils;
 import org.triplea.java.collections.IntegerMap;
-import org.triplea.util.Triple;
 import org.triplea.util.Tuple;
 
 /**
@@ -2491,7 +2487,7 @@ public class UnitAttachment extends DefaultAttachment {
     placementLimit = value;
   }
 
-  private @Nullable Tuple<Integer, String> getPlacementLimit() {
+  public @Nullable Tuple<Integer, String> getPlacementLimit() {
     return placementLimit;
   }
 
@@ -2549,118 +2545,6 @@ public class UnitAttachment extends DefaultAttachment {
 
   public void resetCanRetreatOnStalemate() {
     canRetreatOnStalemate = null;
-  }
-
-  /**
-   * Returns the subset of units that are valid with respect to any stacking limits in effect.
-   *
-   * <p>Note: The passed list of units should have already been filtered for placement restrictions
-   * as otherwise this could return a subset of units that cannot be placed for other reasons.
-   */
-  public static Collection<Unit> filterUnitsByStackingLimit(
-      final Collection<Unit> units,
-      final String limitType,
-      final GamePlayer owner,
-      final Territory t) {
-    final PlayerAttachment pa = PlayerAttachment.get(owner);
-    final Function<UnitAttachment, Tuple<Integer, String>> stackingLimitGetter;
-    final Set<Triple<Integer, String, Set<UnitType>>> playerStackingLimits;
-    switch (limitType) {
-      case "movementLimit":
-        stackingLimitGetter = UnitAttachment::getMovementLimit;
-        playerStackingLimits = (pa == null ? Set.of() : pa.getMovementLimit());
-        break;
-      case "attackingLimit":
-        stackingLimitGetter = UnitAttachment::getAttackingLimit;
-        playerStackingLimits = (pa == null ? Set.of() : pa.getAttackingLimit());
-        break;
-      case "placementLimit":
-        stackingLimitGetter = UnitAttachment::getPlacementLimit;
-        playerStackingLimits = (pa == null ? Set.of() : pa.getPlacementLimit());
-        break;
-      default:
-        throw new IllegalArgumentException("Invalid limitType: " + limitType);
-    }
-
-    // Note: This must check each unit individually and track the ones that passed in order to
-    // correctly handle stacking limits that apply to multiple unit types.
-    final var unitsAllowedSoFar = new ArrayList<Unit>();
-    for (final Unit unit : units) {
-      UnitType ut = unit.getType();
-      Tuple<Integer, String> stackingLimit = stackingLimitGetter.apply(ut.getUnitAttachment());
-      int maxAllowed =
-          getMaximumNumberOfThisUnitTypeToReachStackingLimit(
-              ut, t, owner, stackingLimit, playerStackingLimits, unitsAllowedSoFar);
-      if (maxAllowed > 0) {
-        unitsAllowedSoFar.add(unit);
-      }
-    }
-    return unitsAllowedSoFar;
-  }
-
-  /**
-   * Returns the maximum number of units of the specified type that can be placed in the specified
-   * territory according to the specified stacking limit (movement, attack, or placement).
-   *
-   * @return {@link Integer#MAX_VALUE} if there is no stacking limit for the specified conditions.
-   */
-  private static int getMaximumNumberOfThisUnitTypeToReachStackingLimit(
-      final UnitType ut,
-      final Territory t,
-      final GamePlayer owner,
-      final Tuple<Integer, String> stackingLimit,
-      final Set<Triple<Integer, String, Set<UnitType>>> playerStackingLimits,
-      final Collection<Unit> pendingUnits) {
-    int max = Integer.MAX_VALUE;
-    final UnitAttachment ua = ut.getUnitAttachment();
-    // Concat the territory units with the pending units without copying.
-    final var existingUnits = Iterables.concat(t.getUnits(), pendingUnits);
-    // Apply stacking limits coming from the PlayerAttachment.
-    for (final Triple<Integer, String, Set<UnitType>> limit : playerStackingLimits) {
-      final var unitTypes = limit.getThird();
-      if (!unitTypes.contains(ut)) {
-        continue;
-      }
-      final String stackingType = limit.getSecond();
-      Predicate<Unit> stackingMatch = Matches.unitIsOfTypes(unitTypes);
-      if (stackingType.equals("owned")) {
-        stackingMatch = stackingMatch.and(Matches.unitIsOwnedBy(owner));
-      } else if (stackingType.equals("allied")) {
-        stackingMatch = stackingMatch.and(Matches.alliedUnit(owner));
-      }
-      final int totalInTerritory = CollectionUtils.countMatches(existingUnits, stackingMatch);
-      final Integer limitMax = limit.getFirst();
-      max = Math.min(max, limitMax - totalInTerritory);
-    }
-
-    if (stackingLimit == null) {
-      return max;
-    }
-    max = Math.min(max, stackingLimit.getFirst());
-    // under certain rules (classic rules) there can only be 1 aa gun in a territory.
-    final GameProperties properties = t.getData().getProperties();
-    if (max == Integer.MAX_VALUE
-        && (ua.getIsAaForBombingThisUnitOnly() || ua.getIsAaForCombatOnly())
-        && !(Properties.getWW2V2(properties)
-            || Properties.getWW2V3(properties)
-            || Properties.getMultipleAaPerTerritory(properties))) {
-      max = 1;
-    }
-    final Predicate<Unit> stackingMatch;
-    final String stackingType = stackingLimit.getSecond();
-    switch (stackingType) {
-      case "owned":
-        stackingMatch = Matches.unitIsOfType(ut).and(Matches.unitIsOwnedBy(owner));
-        break;
-      case "allied":
-        stackingMatch = Matches.unitIsOfType(ut).and(Matches.isUnitAllied(owner));
-        break;
-      default:
-        stackingMatch = Matches.unitIsOfType(ut);
-        break;
-    }
-    final int totalInTerritory = CollectionUtils.countMatches(existingUnits, stackingMatch);
-    return Math.max(0, max - totalInTerritory);
   }
 
   @Override
